Reorganize src/dashboard into per-tab dirs
