+++ projects/packages/podcast/src/dashboard/settings/style.scss
@@ -1,0 +1,50 @@
+.podcast__settings {
+	max-inline-size: 680px;
+}
+
+.podcast__cover-control {
+	display: grid;
+	grid-template-columns: 220px 1fr;
+	column-gap: 16px;
+	row-gap: 8px;
+	align-items: start;
+}
+
+.podcast__cover-preview {
+	grid-column: 1;
+	grid-row: 1 / span 2;
+	width: 220px;
+	height: 220px;
+	overflow: hidden;
+	border: 1px solid #dcdcde;
+	border-radius: 4px;
+	background: #f6f7f7;
+	display: flex;
+	align-items: center;
+	justify-content: center;
+
+	img {
+		width: 100%;
+		height: 100%;
+		object-fit: cover;
+	}
+}
+
+.podcast__cover-placeholder {
+	font-size: 12px;
+	color: #757575;
+}
+
+.podcast__cover-actions {
+	grid-column: 2;
+	display: flex;
+	gap: 8px;
+	flex-wrap: wrap;
+}
+
+.podcast__cover-warning {
+	grid-column: 2;
+	margin: 0;
+	font-size: 12px;
+	color: #b26200;
+}
