+++ projects/packages/podcast/src/dashboard/style.scss
@@ -16,13 +16,6 @@ body.jetpack_page_jetpack-podcast {
 	padding-block: 8px;
 }
 
-.podcast__section-header {
-	display: flex;
-	flex-direction: column;
-	gap: 4px;
-	margin-block-end: 24px;
-}
-
 .podcast__section-heading {
 	margin: 0;
 	font-size: 24px;
@@ -31,255 +24,10 @@ body.jetpack_page_jetpack-podcast {
 	color: #1e1e1e;
 }
 
-.podcast__section-description {
-	margin: 0;
-	font-size: 13px;
-	line-height: 1.4;
-	color: #757575;
-}
-
-.podcast__card-title {
-	margin: 0;
-	font-size: 15px;
-	font-weight: 600;
-	color: #1e1e1e;
-}
-
-.podcast__settings {
-	max-inline-size: 680px;
-}
-
+// Used by both Settings ("Finish setting up") and Distribution ("Almost ready
+// to submit") notice lists.
 .podcast__settings-issues {
 	margin: 4px 0 0;
 	padding-inline-start: 18px;
 	list-style: disc;
-}
-
-.podcast__cover-control {
-	display: grid;
-	grid-template-columns: 220px 1fr;
-	column-gap: 16px;
-	row-gap: 8px;
-	align-items: start;
-}
-
-.podcast__cover-preview {
-	grid-column: 1;
-	grid-row: 1 / span 2;
-	width: 220px;
-	height: 220px;
-	overflow: hidden;
-	border: 1px solid #dcdcde;
-	border-radius: 4px;
-	background: #f6f7f7;
-	display: flex;
-	align-items: center;
-	justify-content: center;
-
-	img {
-		width: 100%;
-		height: 100%;
-		object-fit: cover;
-	}
-}
-
-.podcast__cover-placeholder {
-	font-size: 12px;
-	color: #757575;
-}
-
-.podcast__cover-actions {
-	grid-column: 2;
-	display: flex;
-	gap: 8px;
-	flex-wrap: wrap;
-}
-
-.podcast__cover-warning {
-	grid-column: 2;
-	margin: 0;
-	font-size: 12px;
-	color: #b26200;
-}
-
-.podcast__directory-list {
-	list-style: none;
-	margin: 0;
-	padding: 0;
-}
-
-.podcast__directory-row {
-	padding-block: 12px;
-	border-block-end: 1px solid #dcdcde;
-
-	&:last-child {
-		border-block-end: none;
-	}
-}
-
-.podcast__feed-copy {
-	display: flex;
-	align-items: center;
-	gap: 8px;
-}
-
-.podcast__feed-copy-input {
-	flex: 1;
-	min-width: 0;
-	padding: 6px 10px;
-	font-family: ui-monospace, SFMono-Regular, monospace;
-	font-size: 13px;
-	border: 1px solid #dcdcde;
-	border-radius: 4px;
-	background: #f6f7f7;
-	color: #1e1e1e;
-}
-
-.podcast__welcome-hero {
-	padding: 40px;
-	border-radius: 8px;
-	background: linear-gradient(135deg, #f0f6fc 0%, #fdf3fc 100%);
-}
-
-.podcast__welcome-title {
-	margin: 0;
-	font-size: 32px;
-	font-weight: 400;
-	line-height: 1.15;
-	letter-spacing: -0.01em;
-}
-
-.podcast__welcome-hero-copy {
-	max-width: 640px;
-}
-
-.podcast__welcome-benefit-icon {
-	display: inline-flex;
-	align-items: center;
-	color: #1e1e1e;
-
-	svg {
-		inline-size: 24px;
-		block-size: 24px;
-		fill: currentColor;
-	}
-}
-
-.podcast__welcome-steps {
-	list-style: none;
-	margin: 0;
-	padding: 0;
-	display: grid;
-	grid-template-columns: repeat(auto-fit, minmax(min(240px, 100%), 1fr));
-	gap: 24px;
-}
-
-.podcast__welcome-step {
-	display: flex;
-	flex-direction: column;
-	gap: 8px;
-}
-
-.podcast__welcome-step-circle {
-	display: grid;
-	place-items: center;
-	inline-size: 32px;
-	block-size: 32px;
-	border-radius: 50%;
-	background: #2c3338;
-	color: #fff;
-	font-size: 14px;
-	font-weight: 600;
-}
-
-.podcast__episode-thumb {
-	width: 56px;
-	height: 56px;
-	object-fit: cover;
-	border-radius: 4px;
-}
-
-.podcast__episode-thumb--placeholder {
-	background: #f0f0f1;
-	border: 1px dashed #dcdcde;
-}
-
-.podcast__empty-state {
-	padding: 48px 16px;
-	text-align: center;
-	color: #757575;
-}
-
-.podcast__submit-modal .components-modal__content {
-	max-width: 520px;
-}
-
-.podcast__submit-steps {
-	list-style: none;
-	margin: 0;
-	padding: 0;
-}
-
-.podcast__submit-step {
-	padding: 16px;
-	background: #f6f7f7;
-	border-radius: 8px;
-}
-
-.podcast__submit-step-title {
-	margin: 0;
-	font-size: 14px;
-	font-weight: 600;
-	color: #1e1e1e;
-}
-
-.podcast__submit-step-row {
-
-	@media ( max-width: 600px ) {
-		flex-direction: column;
-		align-items: stretch;
-	}
-}
-
-.podcast__submit-step-field {
-	flex: 1;
-	min-width: 0;
-}
-
-.podcast__submit-copy-button {
-	min-inline-size: 8rem;
-	align-self: flex-start;
-	justify-content: flex-start;
-}
-
-.podcast__submit-step-notice {
-	margin-block: 8px 0;
-	margin-inline: 0;
-}
-
-.podcast__submit-step-saved {
-	flex: 1;
-	min-width: 0;
-	padding: 8px 12px;
-	background: #fff;
-	border: 1px solid #ddd;
-	border-radius: 4px;
-}
-
-.podcast__submit-step-saved-icon {
-	flex-shrink: 0;
-	color: #007a3d;
-
-	// Gutenberg's Icon paths use fill="currentColor"; force any stray fill="none"
-	// or hard-coded colors to inherit so the success token actually wins.
-	svg {
-		fill: currentColor;
-	}
-}
-
-.podcast__submit-step-saved-url {
-	overflow: hidden;
-	text-overflow: ellipsis;
-	white-space: nowrap;
-	min-width: 0;
 }
