+++ projects/packages/podcast/src/dashboard/welcome/style.scss
@@ -1,0 +1,56 @@
+.podcast__welcome-hero {
+	padding: 40px;
+	border-radius: 8px;
+	background: linear-gradient(135deg, #f0f6fc 0%, #fdf3fc 100%);
+}
+
+.podcast__welcome-title {
+	margin: 0;
+	font-size: 32px;
+	font-weight: 400;
+	line-height: 1.15;
+	letter-spacing: -0.01em;
+}
+
+.podcast__welcome-hero-copy {
+	max-width: 640px;
+}
+
+.podcast__welcome-benefit-icon {
+	display: inline-flex;
+	align-items: center;
+	color: #1e1e1e;
+
+	svg {
+		inline-size: 24px;
+		block-size: 24px;
+		fill: currentColor;
+	}
+}
+
+.podcast__welcome-steps {
+	list-style: none;
+	margin: 0;
+	padding: 0;
+	display: grid;
+	grid-template-columns: repeat(auto-fit, minmax(min(240px, 100%), 1fr));
+	gap: 24px;
+}
+
+.podcast__welcome-step {
+	display: flex;
+	flex-direction: column;
+	gap: 8px;
+}
+
+.podcast__welcome-step-circle {
+	display: grid;
+	place-items: center;
+	inline-size: 32px;
+	block-size: 32px;
+	border-radius: 50%;
+	background: #2c3338;
+	color: #fff;
+	font-size: 14px;
+	font-weight: 600;
+}
